fix: battery level to %

See merge request polytechnique-montr-al/inf3995/20213/equipe-100/inf3995-simulation!13

+++ src/socket_link.h
@@ -19,7 +19,7 @@ struct Status {
   double kalman_state_x;
   double kalman_state_y;
   double kalman_state_z;
-  double drone_battery_level;
+  int drone_battery_level;
   double range_front;
   double range_back;
   double range_left;
